Rename status syncer package and add status write wrapper
